--- const/const.go
@@ -1,6 +1,0 @@
-package _const
-
-//dotweb const
-const (
-	Version = "1.4.9.4"
-)
